[TextFields] Add label animator (#8382)
This PR adds the class that animates the textfield's label. Part of #6942.

One thing I noticed: the label's animation to the floating position (but not from it) doesn't seem to work on simulators running iOS 9, 10, or 11. It works on simulators running iOS 13 simulators though. It also works on @romoore's iPhone 4s device running iOS ~9.3 or so. I think it's just a simulator issue. Still, a little unsettling... 

Here's a gif of the animation:
![label_animation](https://user-images.githubusercontent.com/8020010/64048558-0b555d00-cb40-11e9-87f5-0c6aa2c7fbb2.gif)